feat: providing experimental APIs
