+++ src/internal/OpenApiTools/store/Operation.ts
@@ -1,23 +1,11 @@
 import { EOL } from "os";
 
-import type { OpenApi } from "../../../types";
+import type { Experimental, OpenApi } from "../../../types";
 
 const httpMethodList = ["get", "put", "post", "delete", "options", "head", "patch", "trace"] as const;
 
-export interface Responses {
-  [statusCode: string]: OpenApi.Response;
-}
-
 export interface State {
-  [operationId: string]: {
-    httpMethod: string;
-    requestUri: string;
-    comment: string | undefined;
-    deprecated: boolean;
-    requestBody?: OpenApi.RequestBody;
-    parameters?: OpenApi.Parameter[];
-    responses: Responses;
-  };
+  [operationId: string]: Experimental.OpenApiOperation;
 }
 
 export const create = (rootSchema: OpenApi.Document): State => {
@@ -39,7 +27,7 @@ export const create = (rootSchema: OpenApi.Document): State => {
         deprecated: !!operation.deprecated,
         requestBody: operation.requestBody as OpenApi.RequestBody,
         parameters: operation.parameters as OpenApi.Parameter[],
-        responses: operation.responses as Responses,
+        responses: operation.responses as Experimental.Responses,
       };
     });
   });
