+++ src/types.ts
@@ -2,3 +2,4 @@ export * as CodeGenerator from "./typedef/CodeGenerator";
 export * as OpenApi from "./typedef/OpenApi";
 export * as OpenApiSchemaParser from "./typedef/OpenApiSchemaParser";
 export * as Validator from "./typedef/Validator";
+export * as Experimental from "./typedef/Experimental";
